chore: moves Identity, Multihain-Accounts and Notifications specs to tests
