+++ e2e/specs/quarantine/create-wallet-account.failing.ts
@@ -2,7 +2,7 @@ import { SmokeAccounts } from '../../tags.js';
 import WalletView from '../../pages/wallet/WalletView.js';
 import AccountListBottomSheet from '../../pages/wallet/AccountListBottomSheet.js';
 import Assertions from '../../../tests/framework/Assertions.js';
-import { withMultichainAccountDetailsV2EnabledFixtures } from '../multichain-accounts/common.js';
+import { withMultichainAccountDetailsV2EnabledFixtures } from '../../../tests/helpers/multichain-accounts/common.js';
 import AccountDetails from '../../pages/MultichainAccounts/AccountDetails.js';
 import AddressList from '../../pages/MultichainAccounts/AddressList.js';
 import { defaultGanacheOptions } from '../../../tests/framework/Constants.js';
