+++ tests/smoke/identity/account-syncing/mock-data.ts
