+++ tests/smoke/identity/account-syncing/multi-srp.spec.ts
@@ -1,28 +1,25 @@
-import { loginToApp } from '../../../viewHelper';
-import TestHelpers from '../../../helpers';
-import WalletView from '../../../pages/wallet/WalletView';
-import AccountListBottomSheet from '../../../pages/wallet/AccountListBottomSheet';
-import Assertions from '../../../../tests/framework/Assertions';
-import { SmokeIdentity } from '../../../tags';
+import { loginToApp } from '../../../../e2e/viewHelper';
+import TestHelpers from '../../../../e2e/helpers';
+import WalletView from '../../../../e2e/pages/wallet/WalletView';
+import AccountListBottomSheet from '../../../../e2e/pages/wallet/AccountListBottomSheet';
+import Assertions from '../../../framework/Assertions';
+import { SmokeIdentity } from '../../../../e2e/tags';
 import { withIdentityFixtures } from '../utils/withIdentityFixtures';
 import { arrangeTestUtils } from '../utils/helpers';
 import {
   UserStorageMockttpControllerEvents,
   UserStorageMockttpController,
 } from '../utils/user-storage/userStorageMockttpController';
-import {
-  goToImportSrp,
-  inputSrp,
-} from '../../../../tests/flows/accounts.flow.ts';
-import ImportSrpView from '../../../pages/importSrp/ImportSrpView';
+import { goToImportSrp, inputSrp } from '../../../flows/accounts.flow';
+import ImportSrpView from '../../../../e2e/pages/importSrp/ImportSrpView';
 import { IDENTITY_TEAM_SEED_PHRASE_2 } from '../utils/constants';
 import { createUserStorageController } from '../utils/mocks';
 import {
   USER_STORAGE_GROUPS_FEATURE_KEY,
   USER_STORAGE_WALLETS_FEATURE_KEY,
 } from '@metamask/account-tree-controller';
-import AccountDetails from '../../../pages/MultichainAccounts/AccountDetails';
-import EditAccountName from '../../../pages/MultichainAccounts/EditAccountName';
+import AccountDetails from '../../../../e2e/pages/MultichainAccounts/AccountDetails';
+import EditAccountName from '../../../../e2e/pages/MultichainAccounts/EditAccountName';
 
 describe(SmokeIdentity('Account syncing - Mutiple SRPs'), () => {
   let sharedUserStorageController: UserStorageMockttpController;
