+++ tests/smoke/identity/contact-syncing/contact-sync-toggle.spec.ts
@@ -1,18 +1,18 @@
-import { loginToApp } from '../../../viewHelper';
-import TestHelpers from '../../../helpers.js';
-import Assertions from '../../../../tests/framework/Assertions.ts';
-import { SmokeIdentity } from '../../../tags.js';
+import { loginToApp } from '../../../../e2e/viewHelper';
+import TestHelpers from '../../../../e2e/helpers.js';
+import Assertions from '../../../framework/Assertions';
+import { SmokeIdentity } from '../../../../e2e/tags.js';
 import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
-import { withIdentityFixtures } from '../utils/withIdentityFixtures.ts';
-import { arrangeTestUtils } from '../utils/helpers.ts';
-import { UserStorageMockttpController } from '../utils/user-storage/userStorageMockttpController.ts';
-import TabBarComponent from '../../../pages/wallet/TabBarComponent';
-import SettingsView from '../../../pages/Settings/SettingsView';
-import BackupAndSyncView from '../../../pages/Settings/BackupAndSyncView';
-import { createUserStorageController } from '../utils/mocks.ts';
-import ContactsView from '../../../pages/Settings/Contacts/ContactsView.ts';
-import AddContactView from '../../../pages/Settings/Contacts/AddContactView.ts';
-import CommonView from '../../../pages/CommonView.ts';
+import { withIdentityFixtures } from '../utils/withIdentityFixtures';
+import { arrangeTestUtils } from '../utils/helpers';
+import { UserStorageMockttpController } from '../utils/user-storage/userStorageMockttpController';
+import TabBarComponent from '../../../../e2e/pages/wallet/TabBarComponent';
+import SettingsView from '../../../../e2e/pages/Settings/SettingsView';
+import BackupAndSyncView from '../../../../e2e/pages/Settings/BackupAndSyncView';
+import { createUserStorageController } from '../utils/mocks';
+import ContactsView from '../../../../e2e/pages/Settings/Contacts/ContactsView';
+import AddContactView from '../../../../e2e/pages/Settings/Contacts/AddContactView';
+import CommonView from '../../../../e2e/pages/CommonView';
 
 describe(SmokeIdentity('Contacts syncing - Settings'), () => {
   let sharedUserStorageController: UserStorageMockttpController;
