+++ tests/smoke/identity/utils/constants.ts
