+++ tests/smoke/identity/utils/user-storage/generateEncryptedData.ts
