+++ tests/smoke/identity/utils/withIdentityFixtures.ts
@@ -1,5 +1,5 @@
-import { withFixtures } from '../../../../tests/framework/fixtures/FixtureHelper';
-import FixtureBuilder from '../../../../tests/framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
 import {
   createUserStorageController,
   setupAccountMockedBalances,
