+++ tests/smoke/multichain-accounts/delete-account.spec.ts
@@ -1,16 +1,16 @@
-import { SmokeWalletPlatform } from '../../tags';
+import { SmokeWalletPlatform } from '../../../e2e/tags';
 import {
   SIMPLE_KEYPAIR_ACCOUNT,
   goToAccountDetails,
   withMultichainAccountDetailsEnabledFixtures,
-} from './common';
-import AccountDetails from '../../pages/MultichainAccounts/AccountDetails';
-import DeleteAccount from '../../pages/MultichainAccounts/DeleteAccount';
-import Assertions from '../../../tests/framework/Assertions';
-import Matchers from '../../../tests/framework/Matchers';
-import WalletView from '../../pages/wallet/WalletView';
-import TestHelpers from '../../helpers';
-import AccountListBottomSheet from '../../pages/wallet/AccountListBottomSheet';
+} from '../../helpers/multichain-accounts/common';
+import AccountDetails from '../../../e2e/pages/MultichainAccounts/AccountDetails';
+import DeleteAccount from '../../../e2e/pages/MultichainAccounts/DeleteAccount';
+import Assertions from '../../framework/Assertions';
+import Matchers from '../../framework/Matchers';
+import WalletView from '../../../e2e/pages/wallet/WalletView';
+import TestHelpers from '../../../e2e/helpers';
+import AccountListBottomSheet from '../../../e2e/pages/wallet/AccountListBottomSheet';
 
 const deleteAccount = async () => {
   await AccountDetails.tapDeleteAccountLink();
