+++ tests/smoke/networks/network-manager2.spec.ts
@@ -1,17 +1,17 @@
-import { SmokeNetworkAbstractions } from '../../tags';
-import { loginToApp, navigateToBrowserView } from '../../viewHelper';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import NetworkManager from '../../pages/wallet/NetworkManager';
+import { SmokeNetworkAbstractions } from '../../../e2e/tags';
+import { loginToApp, navigateToBrowserView } from '../../../e2e/viewHelper';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
+import NetworkManager from '../../../e2e/pages/wallet/NetworkManager';
 import { NetworkToCaipChainId } from '../../../app/components/UI/NetworkMultiSelector/NetworkMultiSelector.constants';
-import Assertions from '../../../tests/framework/Assertions';
-import { DappVariants } from '../../../tests/framework/Constants';
-import TabBarComponent from '../../pages/wallet/TabBarComponent';
-import Browser from '../../pages/Browser/BrowserView';
-import TestDApp from '../../pages/Browser/TestDApp';
-import ConnectedAccountsModal from '../../pages/Browser/ConnectedAccountsModal';
-import ConnectBottomSheet from '../../pages/Browser/ConnectBottomSheet';
-import { CustomNetworks } from '../../../tests/resources/networks.e2e';
+import Assertions from '../../framework/Assertions';
+import { DappVariants } from '../../framework/Constants';
+import TabBarComponent from '../../../e2e/pages/wallet/TabBarComponent';
+import Browser from '../../../e2e/pages/Browser/BrowserView';
+import TestDApp from '../../../e2e/pages/Browser/TestDApp';
+import ConnectedAccountsModal from '../../../e2e/pages/Browser/ConnectedAccountsModal';
+import ConnectBottomSheet from '../../../e2e/pages/Browser/ConnectBottomSheet';
+import { CustomNetworks } from '../../resources/networks.e2e';
 
 const POLYGON = CustomNetworks.Tenderly.Polygon.providerConfig.nickname;
