+++ tests/smoke/notifications/utils/constants.ts
