+++ tests/smoke/notifications/utils/mocks.ts
@@ -27,8 +27,8 @@ import {
 import { getDecodedProxiedURL } from './helpers';
 import { MockttpNotificationTriggerServer } from './mock-notification-trigger-server';
 import { mockAuthServices } from '../../identity/utils/mocks';
-import { setupMockRequest } from '../../../../tests/api-mocking/helpers/mockHelpers';
-import { createLogger } from '../../../../tests/framework/logger';
+import { setupMockRequest } from '../../../api-mocking/helpers/mockHelpers';
+import { createLogger } from '../../../framework/logger';
 
 export const mockListNotificationsResponse = getMockListNotificationsResponse();
 mockListNotificationsResponse.response = [
